feat(user-mgmt): implement role management UI with double validation and <1s session invalidation (#132, #133, #134)
